requirements.txt
openai
streamlit
python-dotenv  # optional, for loading env vars from .env
PyPDF2
